Move Plugins to proper namespace #375
