+++ letsencrypt-win-simple/Plugins/IISPlugin.cs
@@ -1,15 +1,15 @@
-﻿using Microsoft.Web.Administration;
-using Microsoft.Win32;
-using Serilog;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Security.Cryptography.X509Certificates;
 using System.Text.RegularExpressions;
 using LetsEncrypt.ACME.Simple.Configuration;
+using Microsoft.Web.Administration;
+using Microsoft.Win32;
+using Serilog;
 
-namespace LetsEncrypt.ACME.Simple
+namespace LetsEncrypt.ACME.Simple.Plugins
 {
     public class IISPlugin : Plugin
     {
