+++ letsencrypt-win-simple/letsencrypt-win-simple.csproj
@@ -159,12 +159,12 @@
     <Compile Include="Extensions\StringExtensions.cs" />
     <Compile Include="Services\LetsEncryptService.cs" />
     <Compile Include="Extensions\ObjectExtensions.cs" />
-    <Compile Include="Plugin\WebDavPlugin.cs" />
-    <Compile Include="Plugin\IISPlugin.cs" />
-    <Compile Include="Plugin\IISSiteServerPlugin.cs" />
-    <Compile Include="Plugin\FTPPlugin.cs" />
-    <Compile Include="Plugin\Plugin.cs" />
-    <Compile Include="Plugin\ManualPlugin.cs" />
+    <Compile Include="Plugins\WebDavPlugin.cs" />
+    <Compile Include="Plugins\IISPlugin.cs" />
+    <Compile Include="Plugins\IISSiteServerPlugin.cs" />
+    <Compile Include="Plugins\FTPPlugin.cs" />
+    <Compile Include="Plugins\Plugin.cs" />
+    <Compile Include="Plugins\ManualPlugin.cs" />
     <Compile Include="Configuration\Options.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
